Former-commit-id: cc5811321006a1cc7d31a8922fb3f53e2aea2090

+++ api/c/files.c
@@ -98,7 +98,7 @@ int sf_memsize()
     } else {
 	memsize = def;
     }
-    return memsize;
+    return (int)memsize;
 }
 
 off_t sf_filesize (sf_file file)
